style: rename dwt_stm32_delay to dwt_delay
