+++ Src/ltc.c
@@ -9,7 +9,7 @@
 #include "constants.h"
 #include "main.h"
 #include "ntc.h"
-#include "dwt_stm32_delay.h"
+#include "dwt_delay.h"
 #include "math.h"
 #include "stdbool.h"
