+++ Src/main.c
@@ -24,10 +24,8 @@
 /* USER CODE BEGIN Includes */
 #include "bms.h"
 #include "constants.h"
-#include "dwt_stm32_delay.h"
-#include "nextion.h"
-#include "DMA_USART.h"
-#include "nextion_functions.h"
+#include "display.h"
+#include "dwt_delay.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
